+++ src/class/traits/encoding_support.rs
@@ -1,11 +1,11 @@
-use {AnyException, Encoding, Hash, AnyObject};
+use {AnyException, Encoding, Hash, AnyObject, Object};
 
 pub trait EncodingSupport {
     fn encode(&self, enc: Encoding, opts: Option<Hash>) -> Self where Self: Sized;
     fn encoding(&self) -> Encoding;
     fn force_encoding(&mut self, enc: Encoding) -> Result<Self, AnyException> where Self: Sized;
     fn is_valid_encoding(&self) -> bool;
-    fn compatible_with<T>(&self, other: T) -> bool where T: Into<AnyObject>;
-    fn compatible_encoding<T>(&self, other: T) -> AnyObject where T: Into<AnyObject>;
+    fn compatible_with(&self, other: &impl Object) -> bool;
+    fn compatible_encoding(&self, other: &impl Object) -> AnyObject;
 }